requirements.txt
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
pyaudiowpatch>=0.2.12.8
faster-whisper>=1.0.0
llama-cpp-python>=0.2.0
numpy>=1.24.0

